fix: cast LWOOBJID to uint64_t for %llu format specifier in CharacterReputationMigration
Agent-Logs-Url: https://github.com/DarkflameUniverse/DarkflameServer/sessions/63bf6fda-1e52-4ad7-a08c-3535f744ae5d

+++ dDatabase/CharacterReputationMigration.cpp
@@ -17,7 +17,7 @@ uint32_t CharacterReputationMigration::Run() {
 
 		tinyxml2::XMLDocument doc;
 		if (doc.Parse(xmlStr.c_str(), xmlStr.size()) != tinyxml2::XML_SUCCESS) {
-			LOG("Failed to parse XML for character %llu during reputation migration", charId);
+			LOG("Failed to parse XML for character %llu during reputation migration", static_cast<uint64_t>(charId));
 			continue;
 		}
